requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
qrcode==6.1
simplejson==3.17.2
six==1.15.0
Werkzeug==1.0.1
